% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacta.multi.loanbook.plot-package.R
\docType{package}
\name{pacta.multi.loanbook.plot-package}
\alias{pacta.multi.loanbook.plot}
\alias{pacta.multi.loanbook.plot-package}
\title{pacta.multi.loanbook.plot: Tools to Visualize Climate Metrics for Multiple Loanbooks}
\description{
This is an experimental package that contains functions to help plot aggregated PACTA results, with a focus on making it easy to navigate and understand results across a large number of banks and loan books.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RMI-PACTA/pacta.multi.loanbook.plot}
  \item Report bugs at \url{https://github.com/RMI-PACTA/pacta.multi.loanbook.plot/issues}
}

}
\author{
\strong{Maintainer}: Jacob Kastl \email{jacob.kastl@gmail.com}

Authors:
\itemize{
  \item Monika Furdyna \email{monika.furdyna@gmail.com} (\href{https://orcid.org/0000-0002-3728-0646}{ORCID}) [contributor]
}

Other contributors:
\itemize{
  \item RMI \email{PACTA4banks@rmi.org} [copyright holder, funder]
}

}
\keyword{internal}
