% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_scatter_alignment_exposure.R
\name{prep_scatter_alignment_exposure}
\alias{prep_scatter_alignment_exposure}
\title{Prepare data to plot using \code{plot_scatter_alignment_exposure()}}
\usage{
prep_scatter_alignment_exposure(
  data,
  year,
  region,
  scenario,
  category,
  exclude_group_ids = "benchmark"
)
}
\arguments{
\item{data}{data.frame. Holds net aggregated alignment metrics on the loan
book level. Must contain columns: \code{group_id}, \code{scenario}, \code{region},
\code{sector}, \code{year}, \code{exposure_weighted_net_alignment},
\code{sum_loan_size_outstanding}.}

\item{year}{Integer. Year on which \code{data} should be filtered.}

\item{region}{Character. Region to filter \code{data} data frame on.}

\item{scenario}{Character. Scenario to filter \code{data} data frame on.}

\item{category}{Character. Character specifying the variable that contains
the groups by which to analyse the loan books. Usually this will be
\code{"group_id"} unless there is a clearly specified reason to use another
category.}

\item{exclude_group_ids}{Character. Character specifying any values from
\code{category} that should not be included in the analysis. This is useful to
remove benchmarks that are not meant to be compared at the same level.
Defaults to \code{"benchmark"}.}
}
\value{
data.frame
}
\description{
Prepare data to plot using \code{plot_scatter_alignment_exposure()}
}
\examples{
# TODO
}
