% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_scatter_animated.R
\name{prep_scatter_animated}
\alias{prep_scatter_animated}
\title{Prepare data to plot animated scatterplot}
\usage{
prep_scatter_animated(
  data_bopo,
  data_net,
  data_level = c("bank", "company"),
  sector,
  region,
  by_group,
  groups_to_plot = NULL
)
}
\arguments{
\item{data_bopo}{data.frame. Data containing buildout and phaseout alignment
values. Must contain columns: \code{'year'}, \code{'sector'}, \code{'region'},
\code{'direction'} and either \code{'name_abcd'} and \code{'alignment_metric'} or
\code{'exposure_weighted_net_alignment'} plus any column implied by \code{by_group}.}

\item{data_net}{data.frame. Data containing net alignment values. Must
contain columns: \code{by_group}, \code{'year'}, \code{'sector'}, \code{'region'}, \code{'direction'} and
either \code{'name_abcd'} and \code{'alignment_metric'} or \code{'exposure_weighted_net_alignment'}.}

\item{data_level}{Character. Level of the plotted data. Can be \code{'bank'} or
\code{'company'}.}

\item{sector}{Character. Sector to filter data on.}

\item{region}{Character. Region to filter data on.}

\item{by_group}{Character. Vector of length 1. Variable to group by.}

\item{groups_to_plot}{Character vector. Groups to filter on.}
}
\value{
data.frame
}
\description{
Prepare data to plot animated scatterplot
}
\examples{
# TODO
}
